+++ bot/src/api/speedrun.rs
@@ -4,7 +4,7 @@ use crate::{api::RequestBuilder, utils::PtDuration};
 use chrono::{DateTime, NaiveDate, Utc};
 use failure::Error;
 use hashbrown::HashMap;
-use reqwest::{header, r#async::Client, Method, Url};
+use reqwest::{header, r#async::Client, Method, StatusCode, Url};
 use std::collections::BTreeMap;
 
 const V1_URL: &'static str = "https://speedrun.com/api/v1";
@@ -40,10 +40,11 @@ impl Speedrun {
 
     /// Fetch the user by id.
     pub async fn user_by_id(&self, user: String) -> Result<Option<User>, Error> {
-        let data = self
+        let data: Option<Data<User>> = self
             .v1(Method::GET, &["users", user.as_str()])
-            .json_or::<Data<User>>()
+            .json_option(no_content)
             .await?;
+
         Ok(data.map(|d| d.data))
     }
 
@@ -59,15 +60,15 @@ impl Speedrun {
             request = request.query_param("embed", q.as_str());
         }
 
-        let data = request.json_or::<Data<Vec<Run>>>().await?;
+        let data: Option<Data<Vec<Run>>> = request.json_option(no_content).await?;
         Ok(data.map(|d| d.data))
     }
 
     /// Get a game by id.
     pub async fn game_by_id(&self, game: String) -> Result<Option<Game>, Error> {
-        let data = self
+        let data: Option<Data<Game>> = self
             .v1(Method::GET, &["games", game.as_str()])
-            .json_or::<Data<Game>>()
+            .json_option(no_content)
             .await?;
         Ok(data.map(|d| d.data))
     }
@@ -84,14 +85,14 @@ impl Speedrun {
             request = request.query_param("embed", q.as_str());
         }
 
-        let data = request.json_or::<Data<Vec<Category>>>().await?;
+        let data: Option<Data<Vec<Category>>> = request.json_option(no_content).await?;
         Ok(data.map(|d| d.data))
     }
 
     /// Get game levels.
     pub async fn game_levels(&self, game: String) -> Result<Option<Vec<Level>>, Error> {
         let request = self.v1(Method::GET, &["games", game.as_str(), "levels"]);
-        let data = request.json_or::<Data<Vec<Level>>>().await?;
+        let data: Option<Data<Vec<Level>>> = request.json_option(no_content).await?;
         Ok(data.map(|d| d.data))
     }
 
@@ -100,9 +101,9 @@ impl Speedrun {
         &self,
         category: String,
     ) -> Result<Option<Vec<Variable>>, Error> {
-        let data = self
+        let data: Option<Data<Vec<Variable>>> = self
             .v1(Method::GET, &["categories", category.as_str(), "variables"])
-            .json_or::<Data<Vec<Variable>>>()
+            .json_option(no_content)
             .await?;
 
         Ok(data.map(|d| d.data))
@@ -114,10 +115,10 @@ impl Speedrun {
         category: String,
         top: u32,
     ) -> Result<Option<Page<GameRecord>>, Error> {
-        let data = self
+        let data: Option<Page<GameRecord>> = self
             .v1(Method::GET, &["categories", category.as_str(), "records"])
             .query_param("top", top.to_string().as_str())
-            .json_or::<Page<GameRecord>>()
+            .json_option(no_content)
             .await?;
         Ok(data)
     }
@@ -146,8 +147,16 @@ impl Speedrun {
             request = request.query_param(&format!("var-{}", key), &value);
         }
 
-        let data = request.json_or::<Data<GameRecord>>().await?;
+        let data: Option<Data<GameRecord>> = request.json_option(no_content).await?;
         Ok(data.map(|d| d.data))
+    }
+}
+
+/// Handle not found as a missing body.
+fn no_content(status: &StatusCode) -> bool {
+    match *status {
+        StatusCode::NO_CONTENT => true,
+        _ => false,
     }
 }