+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Added `!poll` command to run polls.
+- Added `!weather` command to check the current weather.
 
 [Unreleased]: https://github.com/udoprog/setmod/compare/0.3.7...master